[GameRunner2 Extract Methods] Move 'getFile' method from ClientFileSystemHelper to LoadGameUtil, the one location that needs it
