+++ src/games/strategy/engine/ClientFileSystemHelper.java
@@ -14,7 +14,11 @@ import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.GameRunner2;
 import games.strategy.util.Version;
 
-/** Pure utility class, final and private constructor to enforce this */
+/**
+ * Pure utility class, final and private constructor to enforce this
+ * WARNING: do not call ClientContext.getInstance() in this class. ClientContext call this class in turn
+ * during construction, depending upon ordering this can cause an infinite call loop.
+ */
 public final class ClientFileSystemHelper {
   private ClientFileSystemHelper() {
 
@@ -42,7 +46,7 @@ public final class ClientFileSystemHelper {
   }
 
 
-  private static String getGameRunnerFileLocation(final String runnerClassName) {
+  public static String getGameRunnerFileLocation(final String runnerClassName) {
     final URL url = GameRunner2.class.getResource(runnerClassName);
     String fileName = url.getFile();
 
@@ -57,6 +61,8 @@ public final class ClientFileSystemHelper {
 
 
   private static String getTripleaJarWithEngineVersionStringPath() {
+    // TODO: This is begging for trouble since we call ClientFileSystem during the construction of
+    // ClientContext. Though, we will at this point already have parsed the game engine version, so it is okay (but brittle)
     ClientContext context = ClientContext.getInstance();
     EngineVersion engine = context.engineVersion();
     Version version = engine.getVersion();
@@ -117,70 +123,6 @@ public final class ClientFileSystemHelper {
   }
 
 
-  /**
-   * Search for a file that may be contained in one of multiple folders.
-   *
-   * The file to search for is given by first parameter, second is the list of folders.
-   * We will search all possible paths of the first folder before moving on to the next,
-   * so ordering of the possible folders is more important than the ordering of search paths.
-   *
-   * The search paths vary by if this class is being run from a class file instance,
-   * or a copy compiled into a jar.
-   *
-   * @param game The name of the file to find
-   * @param possibleFolders An array containing a sequence of possible folders that may contain
-   *        the search file.
-   * @return Throws illegal state if not found. Otherwise returns a file reference whose name
-   *         matches the first parameter and parent folder matches an element of "possibleFolders"
-   */
-  public static File getFile(final String game, final String[] possibleFolders) {
-    for (final String possibleFolder : possibleFolders) {
-      final File start = ClientContext.getRootFolder();
-      if (folderContainsFolderAndFile(start, possibleFolder, game)) {
-        return new File(new File(start, possibleFolder), game);
-      }
-
-      final File secondStart = getParentFolder(possibleFolder);
-      if (folderContainsFolderAndFile(secondStart, possibleFolder, game)) {
-        return new File(new File(secondStart, possibleFolder), game);
-      }
-
-    }
-    throw new IllegalStateException(
-        "Could not find any of these folders: " + Arrays.asList(possibleFolders) + ", containing game file: " + game);
-  }
-
-  /* From the Game Runner root location, walk up directories until we find a given folder */
-  private static File getParentFolder(final String folderToFind) {
-    File f = new File(getGameRunnerFileLocation("GameRunner2.class"));
-
-    while (f != null && f.exists() && !folderContains(f, folderToFind)) {
-      f = f.getParentFile();
-    }
-    return f;
-  }
-
-
-  /* Check if a folder contains another folder or file */
-  private static boolean folderContains(final File folder, final String childToFind) {
-    if (folder == null || folder.list() == null || folder.list().length == 0) {
-      return false;
-    }
-    return Arrays.asList(folder.list()).contains(childToFind);
-  }
-
-
-
-  /* Check if a given folder contains another folder that in turn contains a given file */
-  private static boolean folderContainsFolderAndFile(final File f, final String childFolder, final String child) {
-    if (folderContains(f, childFolder)) {
-      final File possibleParent = new File(f, childFolder);
-      if (folderContains(possibleParent, child)) {
-        return true;
-      }
-    }
-    return false;
-  }
 
 
 }
